+++ eventmesh-storage/eventmesh-storage-kafka/src/main/java/org/apache/eventmesh/storage/kafka/producer/KafkaProducerImpl.java
@@ -15,13 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.connector.kafka.producer;
+package org.apache.eventmesh.storage.kafka.producer;
 
 import org.apache.eventmesh.api.RequestReplyCallback;
 import org.apache.eventmesh.api.SendCallback;
 import org.apache.eventmesh.api.producer.Producer;
 import org.apache.eventmesh.common.config.Config;
-import org.apache.eventmesh.connector.kafka.config.ClientConfiguration;
+import org.apache.eventmesh.storage.kafka.config.ClientConfiguration;
 
 import org.apache.kafka.clients.producer.ProducerConfig;
