+++ eventmesh-storage/eventmesh-storage-kafka/src/test/resources/META-INF/services/org.apache.io.openmessaging.producer.Producer
@@ -17,4 +17,4 @@
 # under the License.
 #
 
-org.apache.eventmesh.connector.kafka.producer.ProducerImpl
+org.apache.eventmesh.storage.kafka.producer.ProducerImpl
